+++ turbopack/crates/turbo-tasks-backend/src/kv_backing_storage.rs
@@ -314,8 +314,8 @@ impl TurboBackingStorage {
                 // Byte totals are the physical on-disk bytes (post-compression, including .sst /
                 // .blob / .meta files) produced and removed by the commit.
                 let stats = batch.commit().context("Unable to commit operations")?;
-                snapshot_meta.bytes_written = stats.bytes_written as usize;
-                snapshot_meta.bytes_deleted = stats.bytes_deleted as usize;
+                snapshot_meta.bytes_written = stats.bytes_written;
+                snapshot_meta.bytes_deleted = stats.bytes_deleted;
             }
             Ok(snapshot_meta)
         }
